+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Woodpecker: do not run changelog-check when PR contains `dependabot` label
+- Made the number variable also show placeholders
 
 ## [8.2.2] - 2023-06-28