+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumps `@types/rdfjs__data-model` from 2.0.1 to 2.0.4
 - Bumps `eslint-plugin-ember` from 11.4.6 to 11.9.0
 - Bumps `@typescript-eslint/eslint-plugin` from 5.45.0 to 5.60.1
+- Bumps `ember-velcro` to 2.1.0
 
 ## [8.2.2] - 2023-06-28